feat(bind): add support for attaching event handlers to store changes
This could technically be done manually by directly using the
store.$subscribe API, but this is convenient syntax sugar for it (as
is most of bind.js).

I'm still not sure whether repurposing data-miu-on for it is a good idea
or not, but data-miu-bind makes a bit less sense, and I don't want to
add another attribute for it.