feat(bind): allow specifying bind event trigger, make two-way explicit
This is part of a larger refactor I'm working on, which will make sense
in upcoming commits.

I wanted to allow the flexibility of specifying any event to trigger the
store update, and with that making the default bind be one-way (from
store to element), and making two-way optional with an additional on()
syntax made sense.

The syntax is more complex now, but I think the fact that it's optional,
that it provides additional flexibility, and makes the binding
behavior explicit, makes up for it.