feat(bind): change binding syntax to use arrows
I found the previous colon-separated syntax unintuitive. It didn't make
it clear when a binding was one-way vs two-way.

This version with arrows clearly shows the data flow direction, and
makes this very intuitive. The additional @event syntax simplifies the
previous on(), while still being reasonably comprehensive.

I want to support multiple events soon, and the syntax for that will
likely be @(event1,event2), while keeping the notation for single events
without parenthesis for simplicity.