feat(bind): support event binding in data-miu-bind
This removes the data-miu-on attribute, and allows declaring event
handlers in data-miu-bind. This minimizes the API surface to only two
custom attributes (data-miu-bind and data-miu-for). The syntax is also
updated to align with how events are specified for data binding, for
consistency.

The rationale behind this change is that it improves and streamlines the DX,
and to me as a user it does look better. Claude also agrees. :)