feat: export library version
This uses the Git version when the bundle is built, otherwise it falls
back on the version imported from the package.json. It seems hacky, but
works well from my tests in Chromium.

Browser support for import attributes is decent, except in Firefox,
where it's not supported at all[1]. But since I'm developing exclusively
on Chromium, and the built miu.min.js will not include this (esbuild
tree-shaking works well), I'm fine with keeping it.

[1]: https://caniuse.com/mdn-javascript_statements_import_import_attributes_type_json