feat(bind): add support for reevaluating computed values
This is a quick hack to make it possible for computed values to be
recomputed by updating their store path. Currently computed values are
an afterthought because I've been avoiding the complexity of
implementing them properly, which would require keeping track of any
store path dependencies they use internally. I experimented with this
briefly, but I'm not sure if the added complexity is warranted. With
this change, the user could simply update the computed value manually
when they need to trigger a UI update, which is a bit clunky, but not
too bad considering the alternative.