refactor(bind): generic property vs attribute bind check, parse bind target
This removes the previous hard-coded approach of determining whether to
bind on a property or attribute, in favor of a more generic approach
that checks if a setter exists on the element. This makes it easier to
add support for setting e.g. style properties.

It also extracts some blocks into standandalone functions, and
processing of special bindings like `text` is now better integrated.