feat(store): notify when array length changes
This will also have to support notifying when array elements are removed
using any of the Array methods, but determining which element was
removed is a bit tricky...

Since eventually I would like to pass the subscribed handler not just
the new value, which is currently a useless `undefined` when elements
are removed, but metadata about the change as well. So, for example, the
changeset would contain things like event type, and in the case of array
changes details about the old index, new index, etc.

I still need to think about this feature, and it's not urgent for now,
which is what the TODO is for.