feat(store): wrap primitive values in a StateValue container
This wraps primitive values in the state tree within a StateValue object
with the same interface as StateProxy. This provides more consistent
handling of primitives and creates a unified interface for accessing
values and their metadata. It also simplifies the
SubscriptionManager.notify signature, since it can now extract the path
and root state from all values.

This is a minor breaking change for subscription handling functions,
since in most cases where they would previously receive a primitive
value, they will now receive a StateValue instead. But since the
interface is consistent with StateProxy, it should be easier to maintain
now. The only exception are primitive values of properties on objects
like Map.prototype.size and Array.prototype.length, which were kept as
before in order to maintain a simpler interface.