feat(store): add support for cloning stores with restricted path
This is a hacky PoC to allow reusing the same state for different
stores, while restricting cloned stores to only a subset of the state.
This will be useful for sharing state across web components, for example.

I'm not happy with this implementation, so it will probably have to be
rewritten in the near future.