feat(bind): add support for binding child stores to elements
This builds atop the previous few commits to expose a new data-miu-store
attribute that allows binding new stores to elements that have
restricted access to the parent state. The test here is a contrived
example, but the usefulness will hopefully be more apparent in a more
advanced example I'm currently working on that uses web components.